Update nimbus.guide color schemes to be closer to nimbus.team (#6227)
The new nimbus.team website got launched with different
color schemes than the previous.
This is a small attempt to get the color schemes (dark/light)
closer to those used on nimbus.team.

There was already a dark scheme added before. This change just
sets that dark scheme as default, as is also the case for
nimbus.team.

It also removes orange/amber colors from the light scheme.