Use correct blob subnet when `BLOB_SIDECAR_SUBNET_COUNT_*` is overridden
`BLOB_SIDECAR_SUBNET_COUNT_ELECTRA` is configurable separately from
`MAX_BLOBS_PER_BLOCK_ELECTRA` to allow configs with multiple blob IDs
mapped to the same gossip topic. There is no large runtime overhead in
simply using the correct constant as defined by spec. If we want to
restrict to values that map 1 blob to 1 topic, there should be a check.

Also fix max blobs per block computation to be based on the block slot,
rather than the time at which it was received.