When requesting lists via req/resp, allow runtime config of `maxLen` (#6954)
The maximum number of items that are acceptable for any given req/resp
response can be smaller than the hard maximum. For example, if we only
request blobs for a single block, the limit is `MAX_BLOBS_PER_BLOCK`
so we will eventually reject any response that exceeds that limit even
though it technically fits into the `MAX_REQUEST_BLOB_SIDECARS` based
response type.

This adds an optional `maxResponseItems` parameter to requests that
provide `List` responses, so that the caller can further restrict the
maximum expected number of response chunks. If a peer provides more
chunks than expected, this now fails early with `ResponseChunkOverflow`.