Allow runtime config of `MAX_REQUEST_BLOB_SIDECARS` (#6964)
With `MAX_REQUEST_BLOB_SIDECARS` becoming dynamic, we can no longer use
the value from network config in the `List` datatype and have to rely on
an upper bound constant for the maximum supported value instead and then
check the actual length at runtime.

As a side effect, this also extends `BlobIdentifierList` to use the
correct electra specific request limit instead of the Deneb constant.