ci: set CI_CACHE to runner.temp
When building nim compiler we pass $CI_CACHE to scripts/build_nim.sh as the 4th argument but its not set anywhere.

https://github.com/status-im/nimbus-build-system/blob/e6c2c9da39c2d368d9cf420ac22692e99715d22c/makefiles/targets.mk#L80-L88

This could be the reason for weird caching we have been seeing on runner-node-02-linux-01-eu-hel1

Explicitly setting $CI_CACHE to {{runner.temp}} ensure we use a fresh directory always so that we do not get a cached version of nim commit to build.

https://github.com/status-im/nimbus-build-system/blob/e6c2c9da39c2d368d9cf420ac22692e99715d22c/scripts/build_nim.sh#L214