Streamline block routing from gossip to clearance
In order to avoid unnecessary `Forky`/`Forked` transitions and reduce
the risk for block / blob / column / fork confusion, this PR simplifies
the block processor to predominantly work with `Forky` types and
sidecars typed accordingly.

To achieve this, the `AsyncQueue` is replaced with `AsyncLock` that ends
up working in a similar way for the given use case, ie both maintain a
list of pending work - the advantage of `AsyncLock` is that we don't
have to make extra copies of the block and instead queue the in-flight
future (and its closure context where the block lives) inside the lock.

Apart from the performance and memory usage benefits of not having to
make so many copies along the call chain, this also reduces the risk of
sidecar type contamination across different forks.