Docs: Use consensus light client naming throughout the guide
Use consensus light client naming throughout the nimbus guide
instead of just light client. At the main page this term is
already being used but in the rest of the guide it is not.

It is a source of confusion when naming the page just
"light client" in terms of what this light client is and
what it does.

This will even more so become the case as this guide becomes the
general guide of all Nimbus products (and thus other light clients
such as the verified proxy and the portal client).

So I think it is best to always name it specifically consensus
light client. Other option would be cl light client or beacon
light client (but the latter I like less as I think it is less
valid as it doesn't really do the beacon tasks).