Pass TimeConfig to slotOrZero
Conversion from time to slot also needs access to the TimeConfig.
An alternative design based on storing TimeConfig inside BeaconTime
was explored. While that reduces the amount of refactoring, such a
design wouldn't be as clean, as it suggests to beacon_time module that
each TimeConfig may come from a different BeaconClock. In such a setup
(e.g., a hypothetical multi-network VC with various clocks), one would
have to retain the entire BeaconClock not just the TimeConfig to catch
time operations on times originated from different clocks. Further,
the TimeConfig would have to be passed along internally on every minor
operation such as adding an offset. Therefore, opted to follow existing
approach to just pass TimeConfig explicitly where it will be needed.