Replace `INTERVALS_PER_SLOT` with BPS based slot timings
This enables configuration of:

- `PROPOSER_REORG_CUTOFF_BPS` (only used for REST API and constraints)
- `ATTESTATION_DUE_BPS` (must be ≥ reorg cutoff)
- `AGGREGATE_DUE_BPS` (must be ≥ att cutoff and ≤ 10k)
- `SYNC_MESSAGE_DUE_BPS` (must be == att cutoff)
- `CONTRIBUTION_DUE_BPS` (must be == agg cutoff)

When `INTERVALS_PER_SLOT` was retired, they changed it to be slightly
off, 1/3 became 3333/10000 and 2/3 became 6667/10000, so there is a tiny
bit more time to send attestations, and a tiny bit less for aggregates.

Also updated VC deadlines to be based on the dynamic config.

The constraints are just what Nimbus assumes. The spec does not define
valid ranges of config values.